fix(js bongo): swap out --version for --releaseVersion when sending release notification...
...since it is not getting passed correctly. It is not being handled as a
string, instead it shows up as a boolean (e.g., false). It came up when
Python was trying to release with it. Up until now we have relied on the
version being fetched from the package.json. So this bug appears to have
been there since the beginning. And we don't have an e2e test for the
CLI, so this is also something we didn't test for.